updated solution list
Due to NYT changing the possible solutions from Wordle's original release, a custom solutions list is being uploaded. Dates for known unused words have been set far in the future to enable the main code to incorporate those potential solutions. It is still possible for the solver to fail due to NYT using a word that did not appear in Wordle's original solutions list.